working crc16, error handling, command parsing

+++ common/maveric/src/uart.c
@@ -25,7 +25,7 @@ uint8_t uart_read_byte(uint8_t port) {
 }
 
 void uart_read_buf(uint8_t port, uint8_t* buf, uint8_t len) {
-    size_t i;
+    uint8_t i;
     for (i = 0; i < len; i++) {
         buf[i] = uart_read_byte(port);
     }
@@ -42,7 +42,7 @@ void uart_write_byte(uint8_t port, uint8_t c) {
 }
 
 void uart_write_buf(uint8_t port, uint8_t* buf, uint8_t len) {
-    size_t i;
+    uint8_t i;
     for (i = 0; i < len; i++) {
         uart_write_byte(port, buf[i]);
     }
